page_url: move getStreetUrl()
