+++ assets/scripts/streets/name.js
@@ -11,7 +11,8 @@ export function initStreetNameChangeListener () {
   // We create a string representation of the two values we need to compare
   const select = (state) => JSON.stringify({
     name: state.street.name,
-    creatorId: state.street.creatorId
+    creatorId: state.street.creatorId,
+    namespacedId: state.street.namespacedId
   })
 
   const onChange = (string) => {
